shoes-tranning/training/requirements-training.txt
# Requirements para Treinamento de Stable Diffusion + LoRA
# Otimizado para Apple Silicon (M2 Max)

# PyTorch com suporte MPS
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Hugging Face - Stable Diffusion
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0

# LoRA e Fine-tuning
peft>=0.7.0
bitsandbytes>=0.41.0

# Otimizações (opcional mas recomendado)
xformers>=0.0.23  # Otimizações de atenção

# Datasets e processamento
datasets>=2.15.0
Pillow>=10.0.0

# Utilitários
tqdm>=4.65.0
wandb>=0.16.0  # Tracking de experimentos (opcional)
tensorboard>=2.15.0  # Alternativa ao wandb

# Validação e métricas
scikit-learn>=1.3.0
scipy>=1.11.0

# CLIP para validação
clip @ git+https://github.com/openai/CLIP.git

# Outros
numpy>=1.24.0
einops>=0.7.0
safetensors>=0.4.0
omegaconf>=2.3.0
